Use ConditionAbstractTest for client tests.
This will simplify test maintenance as tests added to  ConditionAbstractTest
will be ran from a clients automatically.

The original ClientConditionTest is renamed to LegacyClientConditionTest.
It should be removed at some point.
